+++ lib/rb_sdl2/event/event_queue.rb
@@ -7,52 +7,40 @@ module RbSDL2
         def clear
           # type が ::SDL2::SDL_DROPFILE, ::SDL2::SDL_DROPTEXT の場合に
           # イベントに含まれる file メンバーのポインターが開放されない。そのため Ruby 側で開放処理を行う。
-          event = Event.new
-          while peep(event, ::SDL2::SDL_GETEVENT,
-                     type: ::SDL2::SDL_DROPFILE..::SDL2::SDL_DROPTEXT) > 0
-            event.clear
+          ptr = ::SDL2::SDL_Event.new[:drop]
+          while peep(ptr, ::SDL2::SDL_GETEVENT, type: ::SDL2::SDL_DROPFILE..::SDL2::SDL_DROPTEXT) > 0
+            ::SDL2::SDL_free(ptr[:file])
           end
-          ::SDL2.SDL_FlushEvents(*EventType.minmax)
+          ::SDL2.SDL_FlushEvents(::SDL2::SDL_FIRSTEVENT, ::SDL2::SDL_LASTEVENT)
         end
 
-        def count(type: nil) = peep(nil, ::SDL2::SDL_PEEKEVENT, type: type)
+        def count = peep(nil, ::SDL2::SDL_PEEKEVENT)
         alias length count
         alias size count
 
         def deq(non_block = false, type: nil)
-          event = Event.new
-          while peep(event, ::SDL2::SDL_GETEVENT, type: type) == 0
-            raise ThreadError, Error.message if non_block
+          event = Event.malloc
+          raise RbSDL2Error if non_block while peep(event, ::SDL2::SDL_GETEVENT, type: type) == 0
+          event
+        end
+
+        def empty? = length == 0
+
+        # non_block に true を与えると、イベントキューが一杯の時に例外が発生する。
+        def enq(event, non_block = false)
+          event_copy(event) do |copy|
+            raise RbSDL2Error if non_block while peep(copy, ::SDL2::SDL_ADDEVENT) == 0
           end
           event
         end
 
-        # ブロックにはイベントキューにあるイベントが渡される。
-        # ブロックへ渡されるイベントはコピーされたものでありブロックの外へ持ち出すことができる。
-        # このメソッドは SDL のイベントキューをロックする。
-        def each = block_given? ? reject! { |event| yield(event.dup); true } : to_enum
-
-        def empty? = length == 0
-
-        # イベントキューが一杯の時に 例外
-        def enq(event, non_block = false)
-          event_copy(event) do |copy|
-            while peep(copy, ::SDL2::SDL_ADDEVENT) == 0
-              raise ThreadError, Error.message if non_block
-            end
-            event
-          end
-        end
-
         private def event_copy(event)
-          copy = event.dup
-          yield(copy).tap do
-            # 成功した場合はメンバーポインターはキューにコピーされた。
-            # ブロックの実行が成功したらメンバーポインターの開放を防ぐ。
-            if event.drop_file? || event.drop_text?
-              copy.to_ptr.autorelease = false
-              ::SDL2.SDL_free(copy)
-            end
+          # SysWMMsg の場合は SDL が内容コピーする。イベントのコピーを行う必要はない。
+          if event.drop_file? || event.drop_text?
+            copy = event.dup
+            yield(copy).tap { copy.to_ptr.__free__ }
+          else
+            yield(event)
           end
         end
 
@@ -65,9 +53,13 @@ module RbSDL2
         private def peep(event, action, type: nil)
           min_type, max_type = case type
                                when nil
-                                 EventType.minmax
+                                 [::SDL2::SDL_FIRSTEVENT, ::SDL2::SDL_LASTEVENT]
                                when Range
-                                 type.minmax
+                                 type.minmax.
+                                   map { |obj| Symbol === obj ? EventType.to_num[obj] : obj }
+                               when Symbol
+                                 num = EventType.to_num[type]
+                                 [num, num]
                                else
                                  [type, type]
                                end
@@ -78,10 +70,9 @@ module RbSDL2
 
         def poll
           main_thread!
-          event = Event.new
+          event = Event.malloc
           ::SDL2::SDL_PollEvent(event).nonzero? && event
         end
-        alias get poll
 
         def pump
           main_thread!
@@ -89,8 +80,8 @@ module RbSDL2
         end
 
         # イベントをキューに入れる。enq との違いは push ではイベントコールバックを起動する。
-        # イベントがキューに入った場合は引数のイベントを戻す。フィルターされた場合は nil を戻す。
-        # キューに入れることを失敗したら例外が発生する。
+        # 成功した場合は引数のイベントを戻す。フィルターされた場合は nil を戻す。
+        # キューに入れることが失敗したら例外が発生する。
         def push(event)
           event_copy(event) do |copy|
             num = ::SDL2.SDL_PushEvent(copy)
@@ -109,20 +100,9 @@ module RbSDL2
           ::SDL2.SDL_HasEvent(::SDL2::SDL_QUIT) == ::SDL2::SDL_TRUE
         end
 
-        require_relative 'event_filter'
-
-        # ブロックへ与えられたイベントは外に持ち出すことはできない。
-        # このメソッドは SDL のイベントキューをロックする。
-        # ブロックの戻り値が false の場合、イベントはキューから取り除かれる。
-        # userdata に渡されたオブジェクトはポインターに変換可能なものである必要がある。
-        def reject!(userdata = nil)
-          func = EventFilter.new { |event| yield(event) || (event.clear; nil) }
-          ::SDL2.SDL_FilterEvents(func, userdata)
-        end
-
         def wait(sec = nil)
           main_thread!
-          event = Event.new
+          event = Event.malloc
           if sec.nil?
             ::SDL2::SDL_WaitEvent(event).nonzero? && event
           elsif sec >= 0